Initial setup to build libsodium as a bazel dependency
The whole purpose of this repository is to be used as a communication
tool with members of the #bazel irc channel or rules_foreign_cc
contributors. I am trying to understand how to use rules_foreign_cc to
build a autotools gmake project. In this case I am trying to build
libsodium. With the current configuration the default toolchain is
being used and an error is reported that the CMakeLists.txt file does
not exist. I am building the project with
TMPDIR=/home/dustfinger/dev/tmp bazel build
--experimental_cc_skylark_api_enabled_packages=@rules_foreign_cc//tools/build_defs,tools/build_defs,@foreign_cc_impl
//src:main --verbose_failures --sandbox_debug. The reason for the
change in TMPDIR path is that my /tmp is not mounted as executable.